RUBYAPI-40 config based
