+++ lib/spark_api.rb
@@ -13,6 +13,7 @@ require 'spark_api/request'
 require 'spark_api/connection'
 require 'spark_api/client'
 require 'spark_api/faraday_middleware'
+require 'spark_api/reso_faraday_middleware'
 require 'spark_api/primary_array'
 require 'spark_api/options_hash'
 require 'spark_api/models'
@@ -39,14 +40,6 @@ module SparkApi
     Thread.current[:spark_api_client] ||= SparkApi::Client.new(opts)
   end
 
-  def self.reso_client(opts={})
-    opts = opts.dup
-    opts[:version] = "Reso/OData"
-    opts[:middleware] = :reso_api
-
-    Thread.current[:reso_api_client] ||= SparkApi::Client.new(opts)
-  end
-
   def self.method_missing(method, *args, &block)
     return super unless (client.respond_to?(method))
     client.send(method, *args, &block)
@@ -55,7 +48,6 @@ module SparkApi
   def self.reset
     reset_configuration
     Thread.current[:spark_api_client] = nil
-    Thread.current[:reso_api_client] = nil
   end
 
 end
