+++ lib/spark_api/faraday_middleware.rb
@@ -87,23 +87,7 @@ module SparkApi
     
   end
 
-  class ResoFaradayMiddleware < FaradayMiddleware
-
-    def on_complete(env)
-
-      body = decompress_body(env)
-      body = MultiJson.decode(body)
-
-      if body["D"]
-        super(env)
-        return
-      end
-
-      env[:body] = body
-    end
-
-  end
-
   Faraday::Response.register_middleware :spark_api => FaradayMiddleware
-  Faraday::Response.register_middleware :reso_api => ResoFaradayMiddleware
+
 end
+
